modify RoomPage FC & add RoomMain FC
